requirements.txt
et-xmlfile==1.1.0
numpy==1.23.5
openpyxl==3.0.10
pandas==1.5.1
psycopg2-binary==2.9.5
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
six==1.16.0